+++ src/pint/models/parameter/__init__.py
@@ -41,317 +41,10 @@ from pint.models.parameter.param_base import (
 )
 from pint.models.parameter.str_parameter import strParameter
 from pint.models.parameter.angle_parameter import AngleParameter
+from pint.models.parameter.pair_parameter import pairParameter
+from pint.models.parameter.prefix_parameter import prefixParameter
 from pint.pulsar_mjd import str2longdouble, time_to_mjd_string
 from pint.toa_select import TOASelect
-from pint.utils import split_prefixed_name
-
-
-class prefixParameter:
-    """Families of parameters identified by a prefix like ``DMX_0123``.
-
-    Creating a ``prefixParameter`` is like creating a normal parameter, except that the
-    name should be in the format of prefix and index. For example, ``DMX_0001`` or
-    ``F22``. Appropriate units will be inferred.
-
-    To create a prefix parameter with the same prefix but different index, just
-    use the :meth:`pint.models.parameter.prefixParameter.new_param` method. It will return a new ``prefixParameter`` with the
-    same setup but a new index. Some  units and descriptions will
-    be changed once the index has been changed. The new parameter will not inherit the ``frozen`` status of its parent by default.  In order to get the right units
-    and description, ``.unit_template`` and ``.description_template`` should be provided. If
-    not the new prefix parameter will use the same units and description with
-    the old one. A typical description and units template is like::
-
-        >>> description_template = lambda x: 'This is the description of parameter %d'%x
-        >>> unit_template = lambda x: 'second^%d'%x
-
-    Although it is best to avoid using lambda functions
-
-    Parameters
-    ----------
-    parameter_type : str, optional
-        Example parameter class template for quantity and value setter
-    name : str optional
-        The name of the parameter. It has to be in the format of prefix + index.
-    value
-        Initial parameter value
-    units : str, optional
-        Units that the value is expressed in
-    unit_template : callable
-        The unit template for prefixed parameter
-    description : str, optional
-        Description for the parameter
-    description_template : callable
-        Description template for prefixed parameters
-    prefix_aliases : list of str, optional
-        Alias for the prefix
-    frozen : bool, optional
-        A flag specifying whether "fitters" should adjust the value of this
-        parameter or leave it fixed.
-    continuous : bool
-        Whether derivatives with respect to this parameter make sense.
-    parameter_type : str, optional
-        Example parameter class template for quantity and value setter
-    long_double : bool, optional
-        Set float type quantity and value in numpy long doubles.
-    time_scale : str, optional
-        Time scale for MJDParameter class.
-    """
-
-    def __init__(
-        self,
-        parameter_type="float",
-        name=None,
-        value=None,
-        units=None,
-        unit_template=None,
-        description=None,
-        description_template=None,
-        uncertainty=None,
-        frozen=True,
-        continuous=True,
-        prefix_aliases=None,
-        long_double=False,
-        unit_scale=False,
-        scale_factor=None,
-        scale_threshold=None,
-        time_scale="utc",
-        **kwargs,
-    ):
-        # Split prefixed name, if the name is not in the prefixed format, error
-        # will be raised
-        self.name = name
-        self.prefix, self.idxfmt, self.index = split_prefixed_name(name)
-        # Type identifier
-        self.type_mapping = {
-            "float": floatParameter,
-            "str": strParameter,
-            "bool": boolParameter,
-            "mjd": MJDParameter,
-            "angle": AngleParameter,
-            "pair": pairParameter,
-        }
-        self.parameter_type = parameter_type
-        try:
-            self.param_class = self.type_mapping[self.parameter_type.lower()]
-        except KeyError as e:
-            raise ValueError(f"Unknown parameter type '{parameter_type}' ") from e
-
-        # Set up other attributes in the wrapper class
-        self.unit_template = unit_template
-        self.description_template = description_template
-        input_units = units
-        input_description = description
-        self.prefix_aliases = [] if prefix_aliases is None else prefix_aliases
-        # set templates, the templates should be a named function and input is
-        # the index of prefix parameter.
-
-        # Set the description and units for the parameter composition.
-        if self.unit_template is not None:
-            real_units = self.unit_template(self.index)
-        else:
-            real_units = input_units
-        if self.description_template is not None:
-            real_description = self.description_template(self.index)
-        else:
-            real_description = input_description
-        aliases = [pa + self.idxfmt for pa in self.prefix_aliases]
-        self.long_double = long_double
-        # initiate parameter class
-        self.param_comp = self.param_class(
-            name=self.name,
-            value=value,
-            units=real_units,
-            description=real_description,
-            uncertainty=uncertainty,
-            frozen=frozen,
-            continuous=continuous,
-            aliases=aliases,
-            long_double=long_double,
-            time_scale=time_scale,
-            unit_scale=unit_scale,
-            scale_factor=scale_factor,
-            scale_threshold=scale_threshold,
-        )
-        self.is_prefix = True
-        self.time_scale = time_scale
-
-    @property
-    def repeatable(self):
-        return self.param_comp.repeatable
-
-    @property
-    def units(self):
-        return self.param_comp.units
-
-    @units.setter
-    def units(self, unt):
-        self.param_comp.units = unt
-
-    @property
-    def quantity(self):
-        return self.param_comp.quantity
-
-    @quantity.setter
-    def quantity(self, qnt):
-        self.param_comp.quantity = qnt
-
-    @property
-    def value(self):
-        return self.param_comp.value
-
-    @value.setter
-    def value(self, val):
-        self.param_comp.value = val
-
-    @property
-    def uncertainty(self):
-        return self.param_comp.uncertainty
-
-    @uncertainty.setter
-    def uncertainty(self, ucty):
-        self.param_comp.uncertainty = ucty
-
-    @property
-    def uncertainty_value(self):
-        return self.param_comp.uncertainty_value
-
-    @uncertainty_value.setter
-    def uncertainty_value(self, val):
-        self.param_comp.uncertainty_value = val
-
-    @property
-    def prior(self):
-        return self.param_comp.prior
-
-    @prior.setter
-    def prior(self, p):
-        self.param_comp.prior = p
-
-    @property
-    def aliases(self):
-        return self.param_comp.aliases
-
-    @aliases.setter
-    def aliases(self, a):
-        self.param_comp.aliases = a
-
-    @property
-    def use_alias(self):
-        return self.param_comp.use_alias
-
-    @use_alias.setter
-    def use_alias(self, a):
-        self.param_comp.use_alias = a
-
-    @property
-    def continuous(self):
-        return self.param_comp.continuous
-
-    @continuous.setter
-    def continuous(self, val):
-        self.param_comp.continuous = val
-
-    @property
-    def frozen(self):
-        return self.param_comp.frozen
-
-    @frozen.setter
-    def frozen(self, val):
-        self.param_comp.frozen = val
-
-    @property
-    def description(self):
-        return self.param_comp.description
-
-    @description.setter
-    def description(self, val):
-        self.param_comp.description = val
-
-    @property
-    def special_arg(self):
-        return self.param_comp.special_arg
-
-    def __repr__(self):
-        return self.param_comp.__repr__()
-
-    def from_parfile_line(self, line):
-        return self.param_comp.from_parfile_line(line)
-
-    def prior_pdf(self, value=None, logpdf=False):
-        return self.param_comp.prior_pdf(value, logpdf)
-
-    def str_quantity(self, quan):
-        return self.param_comp.str_quantity(quan)
-
-    def _print_uncertainty(self, uncertainty):
-        return str(uncertainty.to(self.units).value)
-
-    def name_matches(self, name):
-        return self.param_comp.name_matches(name)
-
-    def as_parfile_line(self, format="pint"):
-        return self.param_comp.as_parfile_line(format=format)
-
-    def help_line(self):
-        return self.param_comp.help_line()
-
-    def prefix_matches(self, prefix):
-        return (prefix == self.prefix) or (prefix in self.prefix_aliases)
-
-    def new_param(self, index, inheritfrozen=False):
-        """Get one prefix parameter with the same type.
-
-        Parameters
-        ----------
-        index : int
-            index of prefixed parameter.
-        inheritfrozen : bool, optional
-            whether or not the parameter should inherit the "frozen" status of the base parameter
-
-        Returns
-        -------
-        A prefixed parameter with the same type of instance.
-        """
-
-        new_name = self.prefix + format(index, f"0{len(self.idxfmt)}")
-        kws = {
-            key: getattr(self, key)
-            for key in [
-                "units",
-                "unit_template",
-                "description",
-                "description_template",
-                "frozen",
-                "continuous",
-                "prefix_aliases",
-                "long_double",
-                "time_scale",
-                "parameter_type",
-            ]
-            if hasattr(self, key) and (key != "frozen" or inheritfrozen)
-        }
-        return prefixParameter(name=new_name, **kws)
-
-    def as_ufloat(self, units=None):
-        """Return the parameter as a :class:`uncertainties.ufloat`
-
-        Will cast to the specified units, or the default
-        If the uncertainty is not set will be returned as 0
-
-        Parameters
-        ----------
-        units : astropy.units.core.Unit, optional
-            Units to cast the value
-
-        Returns
-        -------
-        uncertainties.ufloat
-        """
-        if units is None:
-            units = self.units
-        value = self.quantity.to_value(units) if self.quantity is not None else 0
-        error = self.uncertainty.to_value(units) if self.uncertainty is not None else 0
-        return ufloat(value, error)
 
 
 class maskParameter(floatParameter):
@@ -734,181 +427,6 @@ class maskParameter(floatParameter):
         if self.key != other_param.key:
             return False
         return self.key_value == other_param.key_value
-
-
-class pairParameter(floatParameter):
-    """Parameter type for parameters that need two input floats.
-
-    One example are WAVE parameters.
-
-    Parameters
-    ----------
-    name : str
-        The name of the parameter.
-    value : astropy Time, str, float in mjd, str in mjd.
-        The input parameter MJD value.
-    description : str, optional
-        A short description of what this parameter means.
-    uncertainty : number
-        Current uncertainty of the value.
-    frozen : bool, optional
-        A flag specifying whether "fitters" should adjust the value of this
-        parameter or leave it fixed.
-    continuous : bool, optional, default True
-        A flag specifying whether phase derivatives with respect to this
-        parameter exist.
-    aliases : str, optional
-        List of aliases for the current parameter
-
-    """
-
-    def __init__(
-        self,
-        name=None,
-        index=None,
-        value=None,
-        long_double=False,
-        units=None,
-        description=None,
-        uncertainty=None,
-        frozen=True,
-        continuous=False,
-        aliases=[],
-        **kwargs,
-    ):
-        self.index = index
-        name_param = name
-        self.origin_name = name
-        self.prefix = self.origin_name
-
-        self.prefix_aliases = aliases
-
-        super().__init__(
-            name=name_param,
-            value=value,
-            units=units,
-            description=description,
-            uncertainty=uncertainty,
-            frozen=frozen,
-            continuous=continuous,
-            aliases=aliases,
-            long_double=long_double,
-            **kwargs,
-        )
-
-        self.is_prefix = True
-
-    def name_matches(self, name):
-        if super().name_matches(name):
-            return True
-        name_idx = name + str(self.index)
-        return super().name_matches(name_idx)
-
-    def from_parfile_line(self, line):
-        """Read mask parameter line (e.g. JUMP).
-
-        Notes
-        -----
-        The accepted format:
-            NAME value_a value_b
-
-        """
-        try:
-            k = line.split()
-            name = k[0].upper()
-        except IndexError:
-            return False
-        # Test that name matches
-        if not self.name_matches(name):
-            return False
-        try:
-            self.value = (k[1], k[2])
-        except IndexError:
-            return False
-        if name != self.name:
-            # FIXME: what about prefix/mask parameters?
-            self.use_alias = name
-
-        return True
-
-    def as_parfile_line(self, format="pint"):
-        quantity = self.quantity
-        if self.quantity is None:
-            return ""
-        name = self.name if self.use_alias is None else self.use_alias
-        line = "%-15s " % name
-        line += "%25s" % self.str_quantity(quantity[0])
-        line += " %25s" % self.str_quantity(quantity[1])
-
-        return line + "\n"
-
-    def new_param(self, index):
-        """Create a new but same style mask parameter."""
-        return pairParameter(
-            name=self.origin_name,
-            index=index,
-            long_double=self.long_double,
-            units=self.units,
-            aliases=self.prefix_aliases,
-        )
-
-    def _set_quantity(self, vals):
-        vals = [floatParameter._set_quantity(self, val) for val in vals]
-        return vals
-
-    def _set_uncertainty(self, vals):
-        return self._set_quantity(vals)
-
-    @property
-    def value(self):
-        """Return the pure value of a parameter.
-
-        This value will associate with parameter default value, which is .units attribute.
-        """
-        return None if self._quantity is None else self._get_value(self._quantity)
-
-    @value.setter
-    def value(self, val):
-        """Method to set .value.
-
-        Setting .value attribute will change the .quantity attribute other than .value attribute.
-        """
-        if val is None:
-            if (
-                not isinstance(self.quantity, (str, bool))
-                and self._quantity is not None
-            ):
-                raise ValueError(
-                    "Setting .value to None will lose the parameter value."
-                )
-            else:
-                self.value = val
-        self._quantity = self._set_quantity(val)
-
-    def str_quantity(self, quan):
-        """Return quantity as a string."""
-        try:
-            # Maybe it's a singleton quantity
-            return floatParameter.str_quantity(self, quan)
-        except AttributeError:
-            # Not a quantity, let's hope it's a list of length two?
-            if len(quan) != 2:
-                raise ValueError(f"Don't know how to print this as a pair: {quan}")
-
-        v0 = quan[0].to(self.units).value
-        v1 = quan[1].to(self.units).value
-        if self._long_double:
-            if not isinstance(v0, np.longdouble):
-                raise TypeError(
-                    f"Parameter {self} is supposed to contain long doubles but contains a float"
-                )
-            if not isinstance(v1, np.longdouble):
-                raise TypeError(
-                    f"Parameter {self} is supposed to contain long doubles but contains a float"
-                )
-        quan0 = str(v0)
-        quan1 = str(v1)
-        return f"{quan0} {quan1}"
 
 
 class funcParameter(floatParameter):
